Unify stories' names in storybook
